church_backend/requirements.txt
Django==4.2.23
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
python-dotenv
djangorestframework-simplejwt==5.3.0
gunicorn
whitenoise
dj-database-url
cloudinary
django-cloudinary-storage
django-filter
django-ratelimit
bleach


